refactor: enable feature flag openraft:storage-v2
Separate the implementations of RaftLogStorage and RaftStateMachine.
