+++ Cargo.toml
@@ -123,6 +123,7 @@ sled = { git = "https://github.com/datafuse-extras/sled", tag = "v0.34.7-datafus
 openraft = { git = "https://github.com/drmingdrmer/openraft", tag = "v0.9.0-alpha.8", features = [
     "serde",
     "tracing-log",
+    "storage-v2",
     "loosen-follower-log-revert", # allows removing all data from a follower and restoring from the leader.
 ] }
