+++ src/meta/service/src/meta_service/meta_node.rs
@@ -36,7 +36,6 @@ use databend_common_meta_raft_store::ondisk::DataVersion;
 use databend_common_meta_raft_store::ondisk::DATA_VERSION;
 use databend_common_meta_raft_store::sm_v002::leveled_store::sys_data_api::SysDataApiRO;
 use databend_common_meta_sled_store::openraft;
-use databend_common_meta_sled_store::openraft::storage::Adaptor;
 use databend_common_meta_sled_store::openraft::ChangeMembers;
 use databend_common_meta_stoerr::MetaStorageError;
 use databend_common_meta_types::protobuf::raft_service_client::RaftServiceClient;
@@ -156,8 +155,8 @@ pub struct MetaNodeStatus {
     pub last_seq: u64,
 }
 
-pub type LogStore = Adaptor<TypeConfig, RaftStore>;
-pub type SMStore = Adaptor<TypeConfig, RaftStore>;
+pub type LogStore = RaftStore;
+pub type SMStore = RaftStore;
 
 /// MetaRaft is a implementation of the generic Raft handling meta data R/W.
 pub type MetaRaft = Raft<TypeConfig>;
@@ -205,7 +204,8 @@ impl MetaNodeBuilder {
 
         let net = Network::new(sto.clone());
 
-        let (log_store, sm_store) = Adaptor::new(sto.clone());
+        let log_store = sto.clone();
+        let sm_store = sto.clone();
 
         let raft = MetaRaft::new(node_id, Arc::new(config), net, log_store, sm_store)
             .await
